rename reader -> read (#13)
